Move visitor to own file
