+++ crates/compiler/src/lib.rs
@@ -3,8 +3,9 @@ mod error_listener;
 mod input_manager;
 mod output;
 mod parser;
+pub(crate) mod visitors;
 
 pub mod prelude {
-    pub(crate) use crate::input_manager::*;
     pub use crate::{compiler::*, error_listener::*, output::*, parser::*};
+    pub(crate) use crate::{input_manager::*, visitors};
 }
