+++ crates/compiler/src/visitors.rs
@@ -1,0 +1,3 @@
+pub(crate) mod string_table_generator_visitor;
+
+pub(crate) use string_table_generator_visitor::*;
